requirements.txt
beautifulsoup4==4.6.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
emoji==0.4.5
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lyricsgenius==1.8.6
MarkupSafe==1.1.1
numpy==1.13.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
scikit-learn==0.19.0
scipy==0.19.1
six==1.15.0
spotipy==2.13.0
SQLAlchemy==1.3.18
text-unidecode==1.0
torch==1.0.1
-e git+https://github.com/huggingface/torchMoji@198f7d4e0711a7d3cd01968812af0121c54477f8#egg=torchmoji
urllib3==1.25.9
uWSGI==2.0.19.1
Werkzeug==1.0.1
